+++ delocate/delocating.py
@@ -1,11 +1,12 @@
 """ Routines to copy / relink library dependencies in trees and wheels
 """
 
-from __future__ import division, print_function
+from __future__ import annotations
 
 import functools
 import logging
 import os
+import re
 import shutil
 import warnings
 from os.path import abspath, basename, dirname, exists, realpath, relpath
@@ -17,6 +18,7 @@ from typing import (
     Dict,
     FrozenSet,
     Iterable,
+    Iterator,
     List,
     Mapping,
     Optional,
@@ -25,6 +27,15 @@ from typing import (
     Tuple,
     Union,
 )
+
+from macholib.mach_o import (  # type: ignore[import-untyped]
+    CPU_TYPE_NAMES,
+    LC_BUILD_VERSION,
+    LC_VERSION_MIN_MACOSX,
+)
+from macholib.MachO import MachO  # type: ignore[import-untyped]
+from packaging.utils import parse_wheel_filename
+from packaging.version import Version
 
 from .libsana import (
     _allow_all,
@@ -35,6 +46,7 @@ from .libsana import (
 )
 from .tmpdirs import TemporaryDirectory
 from .tools import (
+    _is_macho_file,
     _remove_absolute_rpaths,
     dir2zip,
     find_package_dirs,
@@ -50,6 +62,8 @@ logger = logging.getLogger(__name__)
 
 # Prefix for install_name_id of copied libraries
 DLC_PREFIX = "/DLC/"
+
+_PLATFORM_REGEXP = re.compile(r"macosx_(\d+)_(\d+)_(\w+)")
 
 
 class DelocationError(Exception):
@@ -578,6 +592,281 @@ def _make_install_name_ids_unique(
         validate_signature(lib)
 
 
+def _get_macos_min_version(dylib_path: Path) -> Iterator[tuple[str, Version]]:
+    """Get the minimum macOS version from a dylib file.
+
+    Parameters
+    ----------
+    dylib_path : Path
+        The path to the dylib file.
+
+    Yields
+    ------
+    str
+        The CPU type.
+    Version
+        The minimum macOS version.
+    """
+    if not _is_macho_file(dylib_path):
+        return
+    for header in MachO(dylib_path).headers:
+        for cmd in header.commands:
+            if cmd[0].cmd == LC_BUILD_VERSION:
+                version = cmd[1].minos
+            elif cmd[0].cmd == LC_VERSION_MIN_MACOSX:
+                version = cmd[1].version
+            else:
+                continue
+            yield (
+                CPU_TYPE_NAMES.get(header.header.cputype, "unknown"),
+                Version(f"{version >> 16 & 0xFF}.{version >> 8 & 0xFF}"),
+            )
+            break
+
+
+def _get_archs_and_version_from_wheel_name(
+    wheel_name: str,
+) -> dict[str, Version]:
+    """
+    Get the architecture and minimum macOS version from the wheel name.
+
+    Parameters
+    ----------
+    wheel_name : str
+        The name of the wheel.
+
+    Returns
+    -------
+    Dict[str, Version]
+        A dictionary containing the architecture and minimum macOS version
+        for each architecture in the wheel name.
+    """
+    platform_tag_set = parse_wheel_filename(wheel_name)[-1]
+    platform_requirements = {}
+    for platform_tag in platform_tag_set:
+        match = _PLATFORM_REGEXP.match(platform_tag.platform)
+        if match is None:
+            raise ValueError(f"Invalid platform tag: {platform_tag.platform}")
+        major, minor, arch = match.groups()
+        platform_requirements[arch] = Version(f"{major}.{minor}")
+    return platform_requirements
+
+
+def _get_problematic_libs(
+    required_version: Optional[Version],
+    version_lib_dict: Dict[Version, List[Path]],
+    arch: str,
+) -> set[tuple[Path, Version]]:
+    """
+    Filter libraries that require more modern macOS
+    version than the provided one.
+
+    Parameters
+    ----------
+    required_version : Version or None
+        The expected minimum macOS version.
+        If None, return an empty set.
+    version_lib_dict : Dict[Version, List[Path]]
+        A dictionary containing mapping from minimum macOS version to libraries
+        that require that version.
+    arch : str
+        The architecture of the libraries. For proper handle arm64 case
+
+    Returns
+    -------
+    set[tuple[Path, Version]]
+        A set of libraries that require a more modern macOS version than the
+        provided one.
+    """
+    if required_version is None:
+        return set()
+    if arch == "arm64" and required_version < Version("11.0"):
+        # All arm64 libraries require macOS at least 11.0,
+        # So even if user provide lower deployment target,
+        # for example, by setting environment variable
+        # MACOSX_DEPLOYMENT_TARGET=10.15
+        # the binaries still will be compatible with 11.0+ only.
+        # So there is no need to check for compatible with older macOS versions
+        required_version = Version("11.0")
+    bad_libraries: set[tuple[Path, Version]] = set()
+    for library_version, libraries in version_lib_dict.items():
+        if library_version > required_version:
+            bad_libraries.update((path, library_version) for path in libraries)
+    return bad_libraries
+
+
+def _calculate_minimum_wheel_name(
+    wheel_name: str,
+    wheel_dir: Path,
+    require_target_macos_version: Optional[Version],
+) -> tuple[str, set[tuple[Path, Version]]]:
+    """
+    Update wheel name platform tag, based on the architecture
+    of the libraries in the wheel and actual platform tag.
+
+    Parameters
+    ----------
+    wheel_name : str
+        The name of the wheel.
+    wheel_dir : Path
+        The directory of the unpacked wheel.
+    require_target_macos_version : Version or None
+        The target macOS version that the wheel should be compatible with.
+
+    Returns
+    -------
+    str
+        The updated wheel name.
+    set[tuple[Path, Version]]
+        A set of libraries that require a more modern macOS version than the
+        provided one.
+    """
+    # get platform tag from wheel name using packaging
+    if wheel_name.endswith("any.whl"):
+        # universal wheel, no need to update the platform tag
+        return wheel_name, set()
+    arch_version = _get_archs_and_version_from_wheel_name(wheel_name)
+    # get the architecture and minimum macOS version from the libraries
+    # in the wheel
+    version_info_dict: Dict[str, Dict[Version, List[Path]]] = {}
+
+    for lib in wheel_dir.glob("**/*"):
+        for arch, version in _get_macos_min_version(lib):
+            version_info_dict.setdefault(arch.lower(), {}).setdefault(
+                version, []
+            ).append(lib)
+    version_dkt = {
+        arch: max(version) for arch, version in version_info_dict.items()
+    }
+
+    problematic_libs: set[tuple[Path, Version]] = set()
+
+    try:
+        for arch, version in list(arch_version.items()):
+            if arch == "universal2":
+                if version_dkt["arm64"] == Version("11.0"):
+                    arch_version["universal2"] = max(
+                        version, version_dkt["x86_64"]
+                    )
+                else:
+                    arch_version["universal2"] = max(
+                        version, version_dkt["arm64"], version_dkt["x86_64"]
+                    )
+                    problematic_libs.update(
+                        _get_problematic_libs(
+                            require_target_macos_version,
+                            version_info_dict["arm64"],
+                            "arm64",
+                        )
+                    )
+                problematic_libs.update(
+                    _get_problematic_libs(
+                        require_target_macos_version,
+                        version_info_dict["x86_64"],
+                        "x86_64",
+                    )
+                )
+            elif arch == "universal":
+                arch_version["universal"] = max(
+                    version, version_dkt["i386"], version_dkt["x86_64"]
+                )
+                problematic_libs.update(
+                    _get_problematic_libs(
+                        require_target_macos_version,
+                        version_info_dict["i386"],
+                        "i386",
+                    ),
+                    _get_problematic_libs(
+                        require_target_macos_version,
+                        version_info_dict["x86_64"],
+                        "x86_64",
+                    ),
+                )
+            else:
+                arch_version[arch] = max(version, version_dkt[arch])
+                problematic_libs.update(
+                    _get_problematic_libs(
+                        require_target_macos_version,
+                        version_info_dict[arch],
+                        arch,
+                    )
+                )
+    except KeyError as e:
+        raise DelocationError(
+            f"Failed to find any binary with the required architecture: {e}"
+        ) from e
+    prefix = wheel_name.rsplit("-", 1)[0]
+    platform_tag = ".".join(
+        f"macosx_{version.major}_{version.minor}_{arch}"
+        for arch, version in arch_version.items()
+    )
+    return f"{prefix}-{platform_tag}.whl", problematic_libs
+
+
+def _check_and_update_wheel_name(
+    wheel_path: Path,
+    wheel_dir: Path,
+    require_target_macos_version: Optional[Version],
+) -> Path:
+    """
+    Based on curren wheel name and binary files in the wheel,
+    determine the minimum platform tag and update the wheel name if needed.
+
+    Parameters
+    ----------
+    wheel_path : Path
+        The path to the wheel.
+    wheel_dir : Path
+        The directory of the unpacked wheel.
+    require_target_macos_version : Version or None
+        The target macOS version that the wheel should be compatible with.
+        If provided and the wheel does not satisfy the target MacOS version,
+        raise an error.
+    """
+    wheel_name = os.path.basename(wheel_path)
+
+    new_name, problematic_files = _calculate_minimum_wheel_name(
+        wheel_name, Path(wheel_dir), require_target_macos_version
+    )
+    if problematic_files:
+        problematic_files_str = "\n".join(
+            f"{lib_path} has a minimum target of {lib_macos_version}"
+            for lib_path, lib_macos_version in problematic_files
+        )
+        raise DelocationError(
+            "Library dependencies do not satisfy target MacOS"
+            f" version {require_target_macos_version}:\n"
+            f"{problematic_files_str}"
+        )
+    if new_name != wheel_name:
+        wheel_path = wheel_path.parent / new_name
+    return wheel_path
+
+
+def _update_wheelfile(wheel_dir: Path, wheel_name: str) -> None:
+    """
+    Update the WHEEL file in the wheel directory with the new platform tag.
+
+    Parameters
+    ----------
+    wheel_dir : Path
+        The directory of the unpacked wheel.
+    wheel_name : str
+        The name of the wheel.
+        Used for determining the new platform tag.
+    """
+    platform_tag_set = parse_wheel_filename(wheel_name)[-1]
+    (file_path,) = wheel_dir.glob("*.dist-info/WHEEL")
+    with file_path.open(encoding="utf-8") as f:
+        lines = f.readlines()
+    with file_path.open("w", encoding="utf-8") as f:
+        for line in lines:
+            if line.startswith("Tag:"):
+                f.write(f"Tag: {'.'.join(str(x) for x in platform_tag_set)}\n")
+            else:
+                f.write(line)
+
+
 def delocate_wheel(
     in_wheel: str,
     out_wheel: Optional[str] = None,
@@ -590,6 +879,7 @@ def delocate_wheel(
     executable_path: Optional[str] = None,
     ignore_missing: bool = False,
     sanitize_rpaths: bool = False,
+    require_target_macos_version: Optional[Version] = None,
 ) -> Dict[str, Dict[str, str]]:
     """Update wheel by copying required libraries to `lib_sdir` in wheel
 
@@ -637,6 +927,8 @@ def delocate_wheel(
         Continue even if missing dependencies are detected.
     sanitize_rpaths : bool, default=False, keyword-only
         If True, absolute paths in rpaths of binaries are removed.
+    require_target_macos_version : None or Version, optional, keyword-only
+        If provided, the minimum macOS version that the wheel should support.
 
     Returns
     -------
@@ -662,6 +954,7 @@ def delocate_wheel(
     else:
         out_wheel = abspath(out_wheel)
     in_place = in_wheel == out_wheel
+    remove_old = in_place
     with TemporaryDirectory() as tmpdir:
         wheel_dir = realpath(pjoin(tmpdir, "wheel"))
         zip2dir(in_wheel, wheel_dir)
@@ -704,8 +997,18 @@ def delocate_wheel(
             install_id_prefix=DLC_PREFIX + relpath(lib_sdir, wheel_dir),
         )
         rewrite_record(wheel_dir)
+        out_wheel_ = Path(out_wheel)
+        out_wheel_fixed = _check_and_update_wheel_name(
+            out_wheel_, Path(wheel_dir), require_target_macos_version
+        )
+        if out_wheel_fixed != out_wheel_:
+            out_wheel_ = out_wheel_fixed
+            in_place = False
+            _update_wheelfile(Path(wheel_dir), out_wheel_.name)
         if len(copied_libs) or not in_place:
-            dir2zip(wheel_dir, out_wheel)
+            if remove_old:
+                os.remove(in_wheel)
+            dir2zip(wheel_dir, out_wheel_)
     return stripped_lib_dict(copied_libs, wheel_dir + os.path.sep)